(#13) refactor: クエリ、コマンド関数の引数で利用するオブジェクトを別のファイルに移動させた。
queries.ts, commands.ts に移動させている。
他の処理は description.ts にて引数として利用するコマンドオブジェクトを型安全にした以外には import 元を変更しただけである。
